requirements.txt
bidict==0.23.1
blinker==1.8.2
boto3==1.34.144
botocore==1.34.144
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
DBUtils==3.1.0
deprecation==2.1.0
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
orjson==3.10.6
packaging==24.1
PyJWT==2.8.0
pymongo==4.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
requests==2.32.3
s3transfer==0.10.2
simple-websocket==1.0.0
six==1.16.0
urllib3==1.26.16
Werkzeug==3.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
